generation-service/requirements.txt
absl-py==1.4.0
accelerate==0.18.0
anyio==3.6.2
appdirs==1.4.4
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
diffusers==0.14.0
docker-pycreds==0.4.0
executing==1.2.0
fastapi==0.95.0
filelock==3.10.7
fire==0.5.0
flatbuffers==23.3.3
fonttools==4.39.3
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lit==16.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mediapipe==0.9.2.1
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.8
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.14.0
pyparsing==3.0.9
pyre-extensions==0.0.23
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
safetensors==0.3.0
scipy==1.10.1
sentry-sdk==1.18.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
sympy==1.11.1
termcolor==2.2.0
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.4
triton==2.0.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
wandb==0.14.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0
xformers==0.0.18
zipp==3.15.0